Use a single settings file but with env vars for config
This commit changes the system used for configuring Django for different
build-types. Rather than using the DJANGO_CONF environmental variable to
specify a settings file, we now use the same settings file for each
build-type but employ environmental variables to configure the settings.

To make this easier for local development, we use django-dotenv which
reads env variables out of a Foreman-style .env file. The idea being to
have stage/prod versions of this file uploaded to S3 for AWS
deployments. We'll have to see how that goes.

One problem still to solve is that using an .env files loses the ability
to use helper function for filepaths. Instead the absolute path needs to
be hard-coded which is a little annoying at the moment.