User-data now pulls config from S3
This commit alters how the user-data script works. It now requires the
EC2 instance to have a 'bucket' tag which stores the S3 bucket URL. The
user-data script downloads all scripts in the $bucket/bootstrap folder
and executes them. It exports the BUCKET_URL variable to the downloaded
bootstrapping scripts can use it.