Use S3 and django-storages for static/media files
This was quite fiddly to get right. You need to ensure the EC2 instance
has a role that allows the appropriate write access to the S3 bucket.
The Docker container will automatically pick up the access and secret
key - there is no need to specify them.

We use custom storage classes to allow prefixes within the bucket for
storage.